Fix Franka pour early pxr import - To be reviewed by Maximilian
Franka pour training could warn or crash because task configuration
resolution imported pxr before SimulationApp launched. The eager MDP
exports loaded runtime action implementations through Articulation and
the USD cloner.

Split action configuration from runtime implementations, use resolvable
class strings, and lazily export MDP symbols so config construction
remains Kit-free.
