+++ source/isaaclab_tasks/changelog.d/franka-pour-early-pxr-import.rst
@@ -1,0 +1,6 @@
+Fixed
+^^^^^
+
+* Fixed Franka pour env-config construction importing USD ``pxr`` before Kit launch by
+  splitting pour MDP action configs into :mod:`isaaclab_tasks.contrib.franka_pour.mdp.actions_cfg`
+  and converting the pour MDP package to lazy exports.
