+++ source/isaaclab_tasks/isaaclab_tasks/contrib/franka_pour/mdp/__init__.pyi
@@ -1,0 +1,164 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "AlignProgress",
+    "ApproachProgress",
+    "CurriculumGripperPositionAction",
+    "CurriculumGripperPositionActionCfg",
+    "CurriculumJointPositionAction",
+    "CurriculumJointPositionActionCfg",
+    "GraspLiftProgress",
+    "HeldDeliveryProgress",
+    "LiftProgress",
+    "NewlyDeliveredParticles",
+    "NewlySpilledParticles",
+    "PourCurriculum",
+    "PourReferenceProgress",
+    "PourResetDatasetCurriculum",
+    "PourResetMixture",
+    "PourTaskProgress",
+    "PourTiltProgress",
+    "RESET_MIXTURE_REGION_NAMES",
+    "RESET_MIXTURE_STAGE_NAMES",
+    "TrajectoryJointPositionAction",
+    "TrajectoryJointPositionActionCfg",
+    "align_command_progress",
+    "align_cup_over_target",
+    "arm_reference_error_obs",
+    "arm_reference_phase_obs",
+    "cup_pose_obs",
+    "cup_to_target_obs",
+    "cup_velocity_obs",
+    "ee_pose_obs",
+    "excessive_spill",
+    "extreme_rigid_state",
+    "finite_joint_velocity_l2",
+    "finger_position_obs",
+    "finger_velocity_obs",
+    "grasp_cup",
+    "grasp_to_tcp_quat_obs",
+    "gripper_contact_obs",
+    "gripper_target_obs",
+    "gripper_width_obs",
+    "held_delivery_history_obs",
+    "immediate_pour_success",
+    "lift_command_progress",
+    "lift_cup",
+    "lost_grasp_dwell_obs",
+    "lost_lifted_grasp",
+    "media_target_distance_tanh",
+    "nonfinite_failure",
+    "nonterminating_stable_pour_success",
+    "particle_fractions_obs",
+    "particle_out_of_bounds",
+    "particle_transfer_obs",
+    "particles_in_source",
+    "particles_in_target",
+    "pour_success_bonus",
+    "pour_target_fraction_obs",
+    "reach_cup",
+    "reset_dataset_difficulty",
+    "reset_pour_scene",
+    "spilled_particles",
+    "stable_pour_success",
+    "success_dwell_obs",
+    "sustained_pour_success",
+    "target_position_c_obs",
+    "target_pose_obs",
+    "tcp_cup_distance_tanh",
+    "tcp_pose_obs",
+    "tcp_to_grasp_obs",
+    "tcp_to_grasp_position_c_obs",
+    "terminal_failure",
+    "tilt_command_progress",
+    "tilt_over_target",
+    "time_remaining_obs",
+    "trajectory_status_obs",
+    "unsuccessful_time_out",
+]
+
+from .actions import (
+    CurriculumGripperPositionAction,
+    CurriculumJointPositionAction,
+    TrajectoryJointPositionAction,
+)
+from .actions_cfg import (
+    CurriculumGripperPositionActionCfg,
+    CurriculumJointPositionActionCfg,
+    TrajectoryJointPositionActionCfg,
+)
+from .curriculums import PourCurriculum
+from .events import reset_pour_scene
+from .observations import (
+    arm_reference_error_obs,
+    arm_reference_phase_obs,
+    cup_pose_obs,
+    cup_to_target_obs,
+    cup_velocity_obs,
+    ee_pose_obs,
+    finger_position_obs,
+    finger_velocity_obs,
+    grasp_to_tcp_quat_obs,
+    gripper_contact_obs,
+    gripper_target_obs,
+    gripper_width_obs,
+    held_delivery_history_obs,
+    lost_grasp_dwell_obs,
+    particle_fractions_obs,
+    particle_transfer_obs,
+    pour_target_fraction_obs,
+    success_dwell_obs,
+    target_position_c_obs,
+    target_pose_obs,
+    tcp_pose_obs,
+    tcp_to_grasp_obs,
+    tcp_to_grasp_position_c_obs,
+    time_remaining_obs,
+    trajectory_status_obs,
+)
+from .reset_dataset import PourResetDatasetCurriculum, reset_dataset_difficulty
+from .reset_mixture import RESET_MIXTURE_REGION_NAMES, RESET_MIXTURE_STAGE_NAMES, PourResetMixture
+from .rewards import (
+    AlignProgress,
+    ApproachProgress,
+    GraspLiftProgress,
+    HeldDeliveryProgress,
+    LiftProgress,
+    NewlyDeliveredParticles,
+    NewlySpilledParticles,
+    PourReferenceProgress,
+    PourTaskProgress,
+    PourTiltProgress,
+    align_command_progress,
+    align_cup_over_target,
+    finite_joint_velocity_l2,
+    grasp_cup,
+    lift_command_progress,
+    lift_cup,
+    media_target_distance_tanh,
+    particles_in_source,
+    particles_in_target,
+    pour_success_bonus,
+    reach_cup,
+    spilled_particles,
+    sustained_pour_success,
+    tcp_cup_distance_tanh,
+    terminal_failure,
+    tilt_command_progress,
+    tilt_over_target,
+)
+from .terminations import (
+    excessive_spill,
+    extreme_rigid_state,
+    immediate_pour_success,
+    lost_lifted_grasp,
+    nonfinite_failure,
+    nonterminating_stable_pour_success,
+    particle_out_of_bounds,
+    stable_pour_success,
+    unsuccessful_time_out,
+)
+from isaaclab.envs.mdp import *
